LayoutTests/platform/glib/imported/w3c/web-platform-tests/mathml/relations/css-styling/attribute-mapping-002-expected.txt


FAIL mathvariant on the math element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the math element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the math element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the annotation element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the annotation element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the annotation element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the annotation-xml element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the annotation-xml element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the annotation-xml element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the maction element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the maction element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the maction element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the menclose element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the menclose element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the menclose element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the merror element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the merror element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the merror element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mfrac element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mfrac element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mfrac element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mi element is mapped to CSS text-transform assert_equals: no attribute expected "math-auto" but got "none"
FAIL scriptlevel on the mi element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mi element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mmultiscripts element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mmultiscripts element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mmultiscripts element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mn element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mn element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mn element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mo element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mo element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mo element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mover element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mover element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mover element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mpadded element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mpadded element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mpadded element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mphantom element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mphantom element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mphantom element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mprescripts element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mprescripts element is mapped to math-depth(...) assert_equals: no attribute expected (number) 1 but got (string) ""
FAIL displaystyle on the mprescripts element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mroot element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mroot element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mroot element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mrow element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mrow element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mrow element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the ms element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the ms element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the ms element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mspace element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mspace element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mspace element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the msqrt element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the msqrt element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the msqrt element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mstyle element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mstyle element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mstyle element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the msub element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the msub element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the msub element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the msubsup element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the msubsup element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the msubsup element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the msup element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the msup element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the msup element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mtable element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mtable element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mtable element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mtd element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mtd element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mtd element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mtext element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mtext element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mtext element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the mtr element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the mtr element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the mtr element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the munder element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the munder element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the munder element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the munderover element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the munderover element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the munderover element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the none element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the none element is mapped to math-depth(...) assert_equals: no attribute expected (number) 1 but got (string) ""
FAIL displaystyle on the none element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
FAIL mathvariant on the semantics element is mapped to CSS text-transform assert_equals: attribute specified expected "math-fraktur" but got "none"
FAIL scriptlevel on the semantics element is mapped to math-depth(...) assert_equals: no attribute expected (number) 0 but got (string) ""
FAIL displaystyle on the semantics element is mapped to CSS math-style assert_equals: no attribute expected "compact" but got ""
 
   
 
